Improve paths handling, oddly Sublime Text only uses '/' as separator, fixes #2
In python the os.path module is usually used to handle paths in a cross-platform
manner. But sublime text seems to handle paths internally with a unix style.
This forces us to use '/' separator on every platform, and prevent using the
os.path module, which only makes things worse.

It should improve things a lot when used with Windows.